requirements.txt
Flask==3.0.2  
Flask-MySQL==1.5.2  
Flask-CORS==4.0.0  
scikit-learn==1.4.1.post1  
numpy==1.26.4  
pandas==2.2.1  
gunicorn==21.2.0  